+++ app/__init__.py
@@ -1,11 +1,15 @@
-import os
-from flask import Flask, render_template, request
-from dotenv import load_dotenv
+from flask import Flask, render_template
 
-load_dotenv()
 app = Flask(__name__)
-
 
 @app.route('/')
 def index():
-    return render_template('index.html', title="MLH Fellow", url=os.getenv("URL"))
+    return render_template('index.html')
+
+@app.route('/aboutme')
+def aboutme():
+    return render_template('aboutme.html')
+
+@app.route('/work')
+def work():
+    return render_template('work.html')
